WIP pipelining example
